fix Browser can't search content from search bar
apktool has a bug in lastest version, which is that "&amp;" is changed to "&amp;amp;"
after decompile a apk. we try to delete useless "amp;" manually, but apktool compile the apk still has the bug.
so we have to revert the change, and avoid to use apktool for fixing the bug.

Change-Id: I1d24fbd078953293faf86b3f1d7af82b0dd32f7c
Signed-off-by: wangwei_b <wangwei_b@xiaomi.com>